requirements.txt
# dependencies
geemap==0.36.2
ipykernel==6.29.5
joblib==1.5.2
matplotlib==3.10.6
numpy==2.3.3
pandas==2.3.2
python-dotenv==1.1.1
pyYAML==6.0.2
scipy==1.16.2
seaborn==0.13.2
scikit-learn==1.7.2
streamlit==1.51.0
tensorflow==2.20.0
tqdm==4.67.1

# ee api
ee
earthengine-api

# local featuretoolkit (ftk) module
-e ./featuretoolkit